requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
google-cloud-aiplatform>=1.63.0
